+++ tencentcloud/cbs/v20170312/models.go
@@ -851,10 +851,10 @@ func (r *DescribeSnapshotsResponse) FromJsonString(s string) error {
 type DetachDisksRequest struct {
 	*tchttp.BaseRequest
 
-	// 将要解挂的云硬盘ID， 通过[DescribeDisks](/document/product/362/16315)接口查询，单次请求最多可解挂10块弹性云盘。
+	// 将要卸载的云硬盘ID， 通过[DescribeDisks](/document/product/362/16315)接口查询，单次请求最多可卸载10块弹性云盘。
 	DiskIds []*string `json:"DiskIds,omitempty" name:"DiskIds" list`
 
-	// 对于非共享型云盘，会忽略该参数；对于共享型云盘，该参数表示要从哪个CVM实例上解挂云盘。
+	// 对于非共享型云盘，会忽略该参数；对于共享型云盘，该参数表示要从哪个CVM实例上卸载云盘。
 	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
 }