+++ tencentcloud/cdn/v20180606/models.go
@@ -425,9 +425,9 @@ type AdvancedCache struct {
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	IgnoreCacheControl *string `json:"IgnoreCacheControl,omitempty" name:"IgnoreCacheControl"`
 
-	// 忽略源站的 Set-Cookie 头部
-	// on：开启
-	// off：关闭
+	// 当源站返回Set-Cookie头部时，节点是否缓存该头部及body
+	// on：开启，不缓存该头部及body
+	// off：关闭，遵循用户自定义的节点缓存规则
 	// 默认为关闭状态
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	IgnoreSetCookie *string `json:"IgnoreSetCookie,omitempty" name:"IgnoreSetCookie"`