+++ tencentcloud/cme/v20191029/client.go
@@ -743,7 +743,9 @@ func NewGenerateVideoSegmentationSchemeByAiResponse() (response *GenerateVideoSe
     return
 }
 
-// 发起视频智能拆条任务，支持智能生成和平精英集锦、王者荣耀集锦、足球集锦、篮球集锦 、人物集锦、新闻拆条等任务。
+// <li>发起视频智能拆条任务，支持智能生成和平精英集锦、王者荣耀集锦、足球集锦、篮球集锦 、人物集锦、新闻拆条等任务。</li>
+// <li>和平精英集锦和王者荣耀集锦根据击杀场景进行拆条，足球集锦和篮球集锦根据进球场景进行拆条，人物集锦根据人物人脸特征进行拆条，新闻拆条根据导播进行拆条。</li>
+// <li>【本接口内测中，暂不建议使用】</li>
 func (c *Client) GenerateVideoSegmentationSchemeByAi(request *GenerateVideoSegmentationSchemeByAiRequest) (response *GenerateVideoSegmentationSchemeByAiResponse, err error) {
     if request == nil {
         request = NewGenerateVideoSegmentationSchemeByAiRequest()